requirement.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mysql-connector==2.1.6
numpy==1.16.1
oauthlib==3.0.1
pandas==0.24.0
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
tweepy==3.7.0
Unidecode==1.0.23
urllib3==1.24.1
Werkzeug==0.14.1
